Refactor ConnectionEditor into subcomponents
